# BackEnd/service-encuesta/src/main/java/dev/miapp/encuesta/encuesta/dto/BasicAccessDTO.java
package dev.miapp.encuesta.encuesta.dto;

public class BasicAccessDTO {

	private String token;

	public String getToken() {
		return token;
	}

	public void setToken(String token) {
		this.token = token;
	}
}