# BackEnd/service-encuesta/src/main/java/dev/miapp/encuesta/encuesta/dto/LoginDTO.java
package dev.miapp.encuesta.encuesta.dto;

public class LoginDTO {

	private String username;
	
	private String password;
	
	public String getUsername() {
		return username;
	}
	public void setUsername(String username) {
		this.username = username;
	}
	public String getPassword() {
		return password;
	}
	public void setPassword(String password) {
		this.password = password;
	}
}