# BackEnd/service-encuesta/src/main/java/dev/miapp/encuesta/encuesta/dto/RenewPasswordFirstDTO.java
package dev.miapp.encuesta.encuesta.dto;

public class RenewPasswordFirstDTO {

	private String username;
	private String password;
	private String session;
	
	public String getUsername() {
		return username;
	}
	public void setUsername(String username) {
		this.username = username;
	}
	public String getPassword() {
		return password;
	}
	public void setPassword(String password) {
		this.password = password;
	}
	public String getSession() {
		return session;
	}
	public void setSession(String session) {
		this.session = session;
	}
}