# BackEnd/service-encuesta/src/main/java/dev/miapp/encuesta/encuesta/dto/UpdatePasswordDTO.java
package dev.miapp.encuesta.encuesta.dto;

public class UpdatePasswordDTO {

	private String token;
	private String username;
	private String oldPassword;
	private String newPassword;
	private String verificationCode;
	
	public String getToken() {
		return token;
	}
	public void setToken(String token) {
		this.token = token;
	}
	public String getOldPassword() {
		return oldPassword;
	}
	public void setOldPassword(String oldPassword) {
		this.oldPassword = oldPassword;
	}
	public String getNewPassword() {
		return newPassword;
	}
	public void setNewPassword(String newPassword) {
		this.newPassword = newPassword;
	}
	public String getUsername() {
		return username;
	}
	public void setUsername(String username) {
		this.username = username;
	}
	public String getVerificationCode() {
		return verificationCode;
	}
	public void setVerificationCode(String verificationCode) {
		this.verificationCode = verificationCode;
	}
}