# BackEnd/service-encuesta/src/main/java/dev/miapp/encuesta/encuesta/entity/Encuesta.java
package dev.miapp.encuesta.encuesta.entity;

import java.util.Date;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.Table;
import javax.persistence.Temporal;
import javax.persistence.TemporalType;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Entity
@Table(name = "tbl_encuesta")
@Data
@AllArgsConstructor @NoArgsConstructor @Builder
public class Encuesta {
	
	@Id
	@GeneratedValue (strategy = GenerationType.IDENTITY)
	private Long id;
	private String name;
	
	@Column(name="lastname")
	private String lastName;
	private int age;
	private String language;
	
	@Column(name = "create_at")
	@Temporal(TemporalType.TIMESTAMP)
	private Date createAt;

}