# CloudFormation/RestApi Encuesta-prod-swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Servicios Rest de Api Encuestas - Trabajo Final - MitoCode",
    "version": "API v1",
    "title": "RestApi Encuesta"
  },
  "host": "1or5zjb22c.execute-api.us-east-1.amazonaws.com",
  "basePath": "/prod",
  "schemes": [
    "https"
  ],
  "paths": {
    "/api/v1/encuestas": {
      "get": {
        "operationId": "listEncuestaByLanguageUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/ArrayOfEncuesta"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "401 response"
          },
          "403": {
            "description": "403 response"
          },
          "404": {
            "description": "404 response"
          },
          "500": {
            "description": "500 response"
          }
        }
      },
      "post": {
        "operationId": "createEncuestaUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Encuesta",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Encuesta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/Encuesta"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "201": {
            "description": "201 response"
          },
          "401": {
            "description": "401 response"
          },
          "403": {
            "description": "403 response"
          },
          "404": {
            "description": "404 response"
          },
          "500": {
            "description": "500 response",
            "schema": {
              "$ref": "#/definitions/Encuesta"
            }
          }
        }
      },
      "options": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              },
              "Access-Control-Allow-Methods": {
                "type": "string"
              },
              "Access-Control-Allow-Headers": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/api/v1/encuestas/{id}": {
      "get": {
        "operationId": "getEncuestaUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/Encuesta"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "401 response"
          },
          "403": {
            "description": "403 response"
          },
          "404": {
            "description": "404 response"
          },
          "500": {
            "description": "500 response"
          }
        }
      },
      "options": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              },
              "Access-Control-Allow-Methods": {
                "type": "string"
              },
              "Access-Control-Allow-Headers": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/api/v1/security/change-password": {
      "post": {
        "operationId": "updatePasswordUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "UpdatePasswordDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdatePasswordDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/RespuestaApi"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "201": {
            "description": "201 response"
          },
          "401": {
            "description": "401 response"
          },
          "403": {
            "description": "403 response"
          },
          "404": {
            "description": "404 response"
          }
        }
      },
      "options": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              },
              "Access-Control-Allow-Methods": {
                "type": "string"
              },
              "Access-Control-Allow-Headers": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/api/v1/security/first-reset-password": {
      "post": {
        "operationId": "renewPasswordFirstUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "RenewPasswordFirstDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RenewPasswordFirstDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/RespuestaApi"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "201": {
            "description": "201 response"
          },
          "401": {
            "description": "401 response"
          },
          "403": {
            "description": "403 response"
          },
          "404": {
            "description": "404 response"
          }
        }
      },
      "options": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              },
              "Access-Control-Allow-Methods": {
                "type": "string"
              },
              "Access-Control-Allow-Headers": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/api/v1/security/login": {
      "post": {
        "operationId": "loginUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "LoginDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/RespuestaApi"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "201": {
            "description": "201 response"
          },
          "401": {
            "description": "401 response"
          },
          "403": {
            "description": "403 response"
          },
          "404": {
            "description": "404 response"
          }
        }
      },
      "options": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              },
              "Access-Control-Allow-Methods": {
                "type": "string"
              },
              "Access-Control-Allow-Headers": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/api/v1/security/refresh-token": {
      "post": {
        "operationId": "refreshTokenUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "BasicAccessDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BasicAccessDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/RespuestaApi"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "201": {
            "description": "201 response"
          },
          "401": {
            "description": "401 response"
          },
          "403": {
            "description": "403 response"
          },
          "404": {
            "description": "404 response"
          }
        }
      },
      "options": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              },
              "Access-Control-Allow-Methods": {
                "type": "string"
              },
              "Access-Control-Allow-Headers": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/api/v1/security/signout": {
      "post": {
        "operationId": "signOutUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "BasicAccessDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BasicAccessDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/RespuestaApi"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "201": {
            "description": "201 response"
          },
          "401": {
            "description": "401 response"
          },
          "403": {
            "description": "403 response"
          },
          "404": {
            "description": "404 response"
          }
        }
      },
      "options": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              },
              "Access-Control-Allow-Methods": {
                "type": "string"
              },
              "Access-Control-Allow-Headers": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/api/v1/security/token": {
      "post": {
        "operationId": "verificarTokenUsingPOST",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/RespuestaApi"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "201": {
            "description": "201 response"
          },
          "401": {
            "description": "401 response"
          },
          "403": {
            "description": "403 response"
          },
          "404": {
            "description": "404 response"
          }
        }
      },
      "options": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              },
              "Access-Control-Allow-Methods": {
                "type": "string"
              },
              "Access-Control-Allow-Headers": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ArrayOfEncuesta": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Encuesta"
      }
    },
    "RespuestaApi": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "body": {
          "type": "object",
          "properties": {}
        },
        "idToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "sessionId": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      },
      "title": "RespuestaApi"
    },
    "BasicAccessDTO": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      },
      "title": "BasicAccessDTO"
    },
    "Encuesta": {
      "type": "object",
      "properties": {
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "createAt": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "language": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "title": "Encuesta"
    },
    "UpdatePasswordDTO": {
      "type": "object",
      "properties": {
        "newPassword": {
          "type": "string"
        },
        "oldPassword": {
          "type": "string"
        },
        "token": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "verificationCode": {
          "type": "string"
        }
      },
      "title": "UpdatePasswordDTO"
    },
    "LoginDTO": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      },
      "title": "LoginDTO"
    },
    "RenewPasswordFirstDTO": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "session": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      },
      "title": "RenewPasswordFirstDTO"
    }
  }
}